+++ src/include/file.h
@@ -11,7 +11,7 @@ struct file_t {
 };
 
 void file_open(struct file_t* file, const char* restrict filepath);
-void file_read_to_uint8_t_buffer(struct file_t* file, uint8_t* restrict buffer, const uint32_t chars_to_read);
+void file_read_to_uint8_t_buffer(struct file_t* file, uint8_t* restrict buffer, const size_t chars_to_read);
 void file_close(const struct file_t* file);
 
 #endif // !FILE_H
